- Fix 32-bit build errors and tests - Add negative test for 32-bit platform. - Change CABI to fail more lazily when running on an unsupported platform. - Change CLinker layouts to be null on unsupported platforms, instead of failing when initializing the class - Added note to CLinker about failure to initialize on unsupported platforms

+++ src/hotspot/share/utilities/globalDefinitions_visCPP.hpp
@@ -77,7 +77,7 @@
 // On Linux NULL is defined as a special type '__null'. Assigning __null to
 // integer variable will cause gcc warning. Use NULL_WORD in places where a
 // pointer is stored as integer value.
-#define NULL_WORD NULL
+#define NULL_WORD (intptr_t) NULL
 
 #ifdef _WIN64
 typedef int64_t ssize_t;
